% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinplots-package.R
\docType{package}
\name{addinplots}
\alias{addinplots}
\alias{addinplots-package}
\title{Addins for Plotting}
\description{
For use by my elementary statistics students.  Provides a GUI for 
plotting functions (cloud plots, ggplot2, etc., for which we 
don't teach the code.Code output goes to the console or source document.
}

