% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplotAddin.R
\name{bwplotAddin}
\alias{bwplotAddin}
\title{Make a \code{lattice} Box-and-Whisker Plot.}
\usage{
bwplotAddin()
}
\description{
Interactively make a \code{lattice} bwplot. The resulting
code will be emitted as a call to \code{lattice::bwplot}.
function.
}
\details{
Here's how you use it:

1. Highlight a symbol naming a \code{data.frame} in your R session,
   e.g. \code{mtcars},
2. Execute this addin and interactively build the plot.

When you're happy with the plot, press Done.  The code for
the plot  will be be placed at the cursor position.
}

