% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityplotAddin.R
\name{densityplotAddin}
\alias{densityplotAddin}
\title{Make a \code{lattice} density plot.}
\usage{
densityplotAddin()
}
\description{
Interactively make a \code{lattice} denisty plot. The resulting
code will be emitted as a call to \code{lattice::densityplot}.
function.
}
\details{
Here's how you use it:

1. Highlight a symbol naming a \code{data.frame} in your R session,
   e.g. \code{mtcars},
2. Execute this addin, to interactively build the plot..

When you're done, the code performing this operation will be emitted
at the cursor position.
}

